\name{acebayes-internal}
\alias{placentalimits2}
\alias{placsolver1.cpp}
\alias{placsolver1cpp}
\alias{nsel4.cpp}
\alias{nsel4cpp}
\alias{naelcpp}
\alias{nael.cpp}
\alias{modcpp}
\alias{mod.cpp}
\alias{erf}
\alias{placentarr}
\alias{placentarq}
\alias{placentaqr}
\alias{placentaqq}
\alias{placentaopt}

\alias{jakstatlimits2}
\alias{jakstatsolver2.cpp}
\alias{jakstatsolver2cpp}
\alias{jakstatnsel4.cpp}
\alias{jakstatnsel4cpp}
\alias{jakstatsig.cpp}
\alias{jakstatsigcpp}
\alias{jakstatnaelcpp}
\alias{jakstatnael.cpp}
\alias{jakstatrr}
\alias{jakstatrq}
\alias{jakstatqr}
\alias{jakstatqq}
\alias{jakstatopt}

\alias{fitzhughlimits2}
\alias{fitzsolver2.cpp}
\alias{fitzsolver2cpp}
\alias{fitzhughnsel.cpp}
\alias{fitzhughnsel4cpp}
\alias{fitzhughsig.cpp}
\alias{fitzhughsigcpp}
\alias{fitzhughnaelcpp}
\alias{fitzhughnael.cpp}
\alias{fitzhughrr}
\alias{fitzhughrq}
\alias{fitzhughqr}
\alias{fitzhughqq}
\alias{fitzopt}

\alias{compsolver2.cpp}
\alias{compsolver2cpp}
\alias{compartmentalnsel.cpp}
\alias{compartmentalnsel4cpp}
\alias{compartmentalsig.cpp}
\alias{compartmentalsigcpp}
\alias{compartmentalnaelcpp}
\alias{compartmentalnael.cpp}
\alias{compartmentalrr}
\alias{compartmentalrq}
\alias{compartmentalqr}
\alias{compartmentalqq}
\alias{compopt}

\alias{placenta}
\alias{jakstat}
\alias{fitzhugh}
\alias{compartmental}
\alias{spi}

\title{
Internal Objects for \code{aceodes}
}
\description{
These are internal objects for the \code{\link{aceodes}} package.
}
\keyword{internal}