\name{aceodes-package}
\alias{aceodes-package}
\alias{aceodes}
\docType{package}
\title{
Bayesian Optimal Design for Ordinary Differential Equation Models
}
\description{
The functions in this package can be used to reproduce the results in Sections 4 and 5 of Overstall et al (2018). They should only be used having read Overstall et al (2018), including the Supplementary Material, and are designed to be used 
in conjunction with the \code{\link[acebayes]{acebayes}} package (Overstall et al, 2017). The \code{\link[acebayes]{acebayes}} package implements the ACE algorithm (Overstall & Woods, 2017) for finding fully Bayesian optimal designs.

The sole purpose of this package is to reproduce the results in Overstall et al (2018) and the package has limited value beyond that goal.
}
\details{
\tabular{ll}{
Package: \tab aceodes\cr
Type: \tab Package\cr
Version: \tab 1.13\cr
Date: \tab 2019-05-01\cr
License: \tab GPL-2\cr
}
}
\author{
Antony M. Overstall, David C. Woods & Ben M. Parker

Maintainer: Antony. M.Overstall \email{A.M.Overstall@soton.ac.uk}
}
\references{
Overstall, A.M. & Woods, D.C. (2017). Bayesian design of experiments using
approximate coordinate exchange. \emph{Technometrics}, \bold{59}, 458-470.

Overstall, A.M., Woods, D.C. & Adamou, M. (2017). acebayes: An R Package for Bayesian Optimal 
Design of Experiments via Approximate Coordinate Exchange. \emph{arXiv:1705.08096}.

Overstall, A.M., Woods, D.C. & Parker, B.M. (2018). Bayesian optimal design for ordinary differential 
equation models with application in biological science. \emph{arXiv:1509.04099}.
}

\keyword{package}

